refactor/#issue58 refactored the code. Refactored decide by creating the class utility and moved all utility functions there. All test regarding utility functions in DecideTest where moved to UtilityTest.
